+++ src/lib/flows/branch-flow.js
@@ -9,11 +9,15 @@ const {
 } = require("@kie/build-chain-configuration-reader");
 const core = require("@actions/core");
 const { logger } = require("../common");
-const { printCheckoutInformation } = require("../summary");
+const {
+  printCheckoutInformation,
+  printExecutionSummary
+} = require("../summary");
 const {
   executeBuild,
   executeBuildSpecificCommand
 } = require("./common/common-helper");
+const { isError } = require("../util/js-util");
 
 const { execute: executePre } = require("./sections/pre");
 const { execute: executePost } = require("./sections/post");
@@ -47,8 +51,17 @@ async function start(context, options = { skipExecution: false }) {
         readerOptions
       )
     : getTree(context.config.github.inputs.definitionFile, readerOptions);
-  logger.debug("branch-flow.js definitionTree", definitionTree);
-
+  logger.debug(
+    "branch-flow.js definitionTree",
+    definitionTree,
+    context.config.github.inputs.definitionFile,
+    context.config.github.inputs.startingProject
+  );
+  if ([null, undefined].includes(definitionTree)) {
+    throw new Error(
+      `The definition tree is undefined. Does the project ${context.config.github.inputs.startingProject} exist into the definition file ${context.config.github.inputs.definitionFile}?`
+    );
+  }
   let nodeChain = await parentChainFromNode(definitionTree);
   logger.debug("branch-flow.js nodeChain", nodeChain);
 
@@ -85,20 +98,24 @@ async function start(context, options = { skipExecution: false }) {
           context.config.github.repository,
           options
         )
-          .then(() => true)
+          .then(e => e)
           .catch(e => e);
 
     await executePost(
       context.config.github.inputs.definitionFile,
-      executionResult,
+      !isError(executionResult),
       readerOptions
     );
 
-    if (executionResult !== true) {
+    if (isError(executionResult)) {
       logger.error(executionResult);
       throw new Error(
         `Command executions have failed, please review latest execution ${executionResult}`
       );
+    } else {
+      core.startGroup(`[Branch Flow] Execution Summary...`);
+      printExecutionSummary(executionResult);
+      core.endGroup();
     }
   } else {
     logger.info("Execution has been skipped.");