+++ src/lib/flows/single-flow.js
@@ -4,12 +4,16 @@ const {
 } = require("./common/build-chain-flow-helper");
 const { executeBuild } = require("./common/common-helper");
 const { getTreeForProject } = require("@kie/build-chain-configuration-reader");
-const { printCheckoutInformation } = require("../summary");
+const {
+  printCheckoutInformation,
+  printExecutionSummary
+} = require("../summary");
 const { logger } = require("../common");
 const core = require("@actions/core");
 const {
   archiveArtifacts
 } = require("../artifacts/build-chain-flow-archive-artifact-helper");
+const { isError } = require("../util/js-util");
 
 const { execute: executePre } = require("./sections/pre");
 const { execute: executePost } = require("./sections/post");
@@ -50,6 +54,11 @@ async function start(
     readerOptions
   );
   logger.debug("single-flow.js definitionTree", definitionTree);
+  if ([null, undefined].includes(definitionTree)) {
+    throw new Error(
+      `The definition tree is undefined. Does the project ${projectTriggeringJob} exist into the definition file ${context.config.github.inputs.definitionFile}?`
+    );
+  }
 
   const nodeChain = [definitionTree];
 
@@ -77,7 +86,7 @@ async function start(
       projectTriggeringJob,
       options
     )
-      .then(() => true)
+      .then(e => e)
       .catch(e => e);
 
     if (options.isArchiveArtifacts) {
@@ -85,7 +94,7 @@ async function start(
       await archiveArtifacts(
         nodeChain.find(node => node.project === projectTriggeringJob),
         nodeChain,
-        executionResult === true ? ["success", "always"] : ["failure", "always"]
+        isError(executionResult) ? ["failure", "always"] : ["success", "always"]
       );
       core.endGroup();
     } else {
@@ -94,15 +103,19 @@ async function start(
 
     await executePost(
       context.config.github.inputs.definitionFile,
-      executionResult,
+      !isError(executionResult),
       readerOptions
     );
 
-    if (executionResult !== true) {
+    if (isError(executionResult)) {
       logger.error(executionResult);
       throw new Error(
         `Command executions have failed, please review latest execution ${executionResult}`
       );
+    } else {
+      core.startGroup(`[Single Flow] Execution Summary...`);
+      printExecutionSummary(executionResult);
+      core.endGroup();
     }
   } else {
     logger.info("Execution has been skipped.");